# docs/openapi_donations.yaml
...
openapi: 3.1.0
info:
  title: GoodPlay Donations API - GOO-15
  description: |
    **GOO-15 Donation Processing Engine** - Complete virtual wallet and donation system.

    ## Features
    - Virtual wallet management with credit conversion
    - Payment processing with multiple providers
    - Donation workflow with compliance checks
    - Batch processing for high-volume operations
    - Financial analytics and reporting
    - Receipt generation and tax compliance
    - Transaction reconciliation

    ## Endpoints Structure
    - `/api/wallet/*` - Virtual wallet operations
    - `/api/donations/*` - Donation processing
    - `/api/conversion-rates/*` - Credit conversion rates
    - `/api/payments/*` - Payment processing
    - `/api/batch/*` - Batch operations (admin)
    - `/api/compliance/*` - Compliance management (admin)
    - `/api/admin/financial/*` - Financial analytics (admin)

  version: 1.0.0
paths:
...
  /api/wallet:
    get:
      tags:
        - Virtual Wallet
      summary: Get user wallet information
      description: Retrieve wallet balance, statistics, and auto-donation settings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wallet information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: WALLET_RETRIEVED_SUCCESS
                  data:
                    $ref: '#/components/schemas/WalletInfo'
              examples:
                success:
                  summary: Successful wallet retrieval
                  value:
                    success: true
                    message: WALLET_RETRIEVED_SUCCESS
                    data:
                      user_id: user123
                      balance: 25.50
                      total_earned: 125.75
                      total_donated: 100.25
                      auto_donation_enabled: true
                      auto_donation_threshold: 50.0
                      auto_donation_percentage: 25.0
  /api/wallet/transactions:
    get:
      tags:
        - Virtual Wallet
      summary: Get wallet transaction history
      description: Retrieve paginated transaction history for the user's wallet
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: transaction_type
          in: query
          schema:
            type: string
            enum: [earned, donated, bonus, refund]
      responses:
        '200':
          description: Transaction history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: TRANSACTIONS_RETRIEVED_SUCCESS
                  data:
                    type: object
                    properties:
                      transactions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
...
  /api/donations/create:
    post:
      tags:
        - Donations
      summary: Create a new donation
      description: Process a donation from wallet balance or external payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - onlus_id
                - amount
              properties:
                onlus_id:
                  type: string
                  description: Target ONLUS identifier
                amount:
                  type: number
                  format: float
                  minimum: 1.0
                payment_method:
                  type: string
                  enum: [wallet, stripe, paypal, bank_transfer]
                  default: wallet
                message:
                  type: string
                  description: Optional message to ONLUS
      responses:
        '200':
          description: Donation created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: DONATION_CREATED_SUCCESS
                  data:
                    $ref: '#/components/schemas/DonationResponse'
  /api/donations/history:
    get:
      tags:
        - Donations
      summary: Get donation history
      description: Retrieve user's donation history with filtering and pagination
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: onlus_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Donation history retrieved successfully
...
  /api/donations/{donation_id}/receipt:
    get:
      tags:
        - Donations
      summary: Download donation receipt
      description: Generate and download PDF receipt for tax deductibility
      security:
        - bearerAuth: []
      parameters:
        - name: donation_id
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [pdf, json]
            default: pdf
      responses:
        '200':
          description: Receipt generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: RECEIPT_GENERATED_SUCCESS
                  data:
                    $ref: '#/components/schemas/ReceiptData'

  # Conversion Rates Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminRole:
      type: apiKey
      in: header
      name: X-Admin-Role
  schemas:
    WalletInfo:
      type: object
      properties:
        user_id:
          type: string
        balance:
          type: number
          format: float
        total_earned:
          type: number
          format: float
        total_donated:
          type: number
          format: float
        auto_donation_enabled:
          type: boolean
        auto_donation_threshold:
          type: number
          format: float
        auto_donation_percentage:
          type: number
          format: float
        preferred_onlus_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Transaction:
      type: object
      properties:
        transaction_id:
          type: string
        user_id:
          type: string
        transaction_type:
          type: string
          enum: [earned, donated, bonus, refund, adjustment, fee]
        amount:
          type: number
          format: float
        status:
          type: string
          enum: [pending, completed, failed, cancelled, refunded]
        description:
          type: string
        source_type:
          type: string
        onlus_id:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    DonationResponse:
      type: object
      properties:
        donation_id:
          type: string
        onlus_id:
          type: string
        amount:
          type: number
          format: float
        status:
          type: string
        payment_method:
          type: string
        created_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
    ReceiptData:
      type: object
      properties:
        receipt_number:
          type: string
        donation_id:
          type: string
        donor_name:
          type: string
        donor_tax_code:
          type: string
        onlus_name:
          type: string
        onlus_tax_code:
          type: string
        amount:
          type: number
          format: float
        tax_deductible_amount:
          type: number
          format: float
        tax_year:
          type: integer
        issued_date:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
        total_items:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
...